Move islands into the /components folder
